make Production ready special backend

+++ src/main/java/com/doctors/entities/Customers.java
@@ -1,11 +1,16 @@
 package com.doctors.entities;
 
 
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Data;
 
@@ -25,7 +30,7 @@ public class Customers {
 	private String email;
 	
 	@Column(name = "phone")
-	private int phone;
+	private long phone;
 	
 	@Column(name = "city")
 	private String city;
@@ -33,7 +38,7 @@ public class Customers {
 	@Column(name = "password")
 	private String password;
 	
-//	@OneToMany(mappedBy = "customer")
+//	@OneToMany(mappedBy = "customer", cascade = CascadeType.ALL,fetch = FetchType.EAGER)
 //	private List<Test> tests;
